shamu: upate prebuilt kernel
e2a2033 kgsl: do not vmap/memset to zero-out pages

comment #1 in b/18404900 on build LMX37E shows a Shamu with vmap failures
caused by kgsl "fast-path" allocations.  I've profiled these pretty
extensively, and for the 70-80% of the allocation sizes, the "slow path" in the
kernel is actually faster.  This prebuilt switches to that "slow path", and in
the course of doing so eliminates false alarms in kernel messages.  It also
reduces the risk that there won't be sufficient vmalloc space for allocations
that cannot afford to fail.

Change-Id: I7d197b9438919ed0bf3e558ff8acbd679d38518a
Signed-off-by: Iliyan Malchev <malchev@google.com>